Fix data race on older Linux kernels
Dereferencing fname is not enough to copy it to the stack, it needs
an explicit clone.

Before this fix multiple threads calling the mkstemp would have a
data race on the static storage of fname. Which would pass the wrong
data to following unlink. Leading to leaked files in /tmp/, also a
file could be unlinked more than once.

This also removes some unneeded mutability